rename useSingleEditFieldProps to makeSingleEditFieldProps
