+++ protocol-designer/src/components/StepEditForm/StepEditFormComponent.js
@@ -16,7 +16,7 @@ import { ButtonRow } from './ButtonRow'
 import formStyles from '../forms/forms.css'
 import styles from './StepEditForm.css'
 import type { FormData, StepType } from '../../form-types'
-import type { FieldPropsByName } from './fields/useSingleEditFieldProps'
+import type { FieldPropsByName } from './fields/makeSingleEditFieldProps'
 import type { StepFormProps, FocusHandlers } from './types'
 
 const STEP_FORM_MAP: { [StepType]: ?React.ComponentType<StepFormProps> } = {
