+++ protocol-designer/src/components/StepEditForm/fields/StepFormDropdownField.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import { DropdownField, type Options } from '@opentrons/components'
 import cx from 'classnames'
 import type { StepFieldName } from '../../../steplist/fieldLevel'
-import type { FieldProps } from './useSingleEditFieldProps'
+import type { FieldProps } from './makeSingleEditFieldProps'
 import styles from '../StepEditForm.css'
 
 export type StepFormDropdownProps = {
