+++ protocol-designer/src/components/StepEditForm/forms/ThermocyclerForm/ProfileSettings.js
@@ -7,7 +7,7 @@ import { TextField } from '../../fields'
 
 import styles from '../../StepEditForm.css'
 
-import type { FieldPropsByName } from '../../fields/useSingleEditFieldProps'
+import type { FieldPropsByName } from '../../fields/makeSingleEditFieldProps'
 
 type Props = {| propsForFields: FieldPropsByName |}
